Rename folder from button-subtle-ai to button-ai
