+++ demo/components/button-ai/index.html
@@ -4,7 +4,7 @@
 		<link rel="stylesheet" href="/node_modules/@brightspace-ui/core/components/demo/styles.css" type="text/css">
 		<script type="module">
 			import '@brightspace-ui/core/components/demo/demo-page.js';
-			import '../../../src/components/button-subtle-ai/button-subtle-ai.js';
+			import '../../../src/components/button-ai/button-subtle-ai.js';
 		</script>
 		<title>d2l-labs-button-subtle-ai</title>
 		<meta name="viewport" content="width=device-width, initial-scale=1.0">
@@ -12,7 +12,7 @@
 	</head>
 	<body>
 		<d2l-demo-page page-title="d2l-labs-button-subtle-ai">
-			<h2>AI Button (focus)</h2>
+			<h2>Button Subtle AI (focus)</h2>
 			<d2l-demo-snippet>
 				<button id="setFocusButton" style="display: block; margin-block-end: 1rem;">Focus It!</button>
 				<d2l-labs-button-subtle-ai text="New AI Feature"></d2l-labs-button-subtle-ai>
