+++ index.html
@@ -32,7 +32,12 @@
 				</ul>
 			</li>
 			<li><a href="demo/components/attribute-picker/index.html">Attribute Picker</a></li>
-			<li><a href="demo/components/button-subtle-ai/index.html">Button Subtle AI</a></li>
+			<li>
+				Button AI
+				<ul>
+					<a href="demo/components/button-ai/index.html">Button Subtle AI</a>
+				</ul>
+			</li>
 			<li><a href="demo/components/card-overlay/index.html">Card Overlay</a></li>
 			<li><a href="demo/components/checkbox-drawer/index.html">Checkbox Drawer</a></li>
 			<li><a href="demo/components/community/index.html">Community</a></li>
