+++ package.json
@@ -8,7 +8,7 @@
     "./components/accordion.js": "./src/components/accordion/accordion.js",
     "./components/accordion-collapse.js": "./src/components/accordion/accordion-collapse.js",
     "./components/attribute-picker.js": "./src/components/attribute-picker/attribute-picker.js",
-    "./components/button-subtle-ai.js": "./src/components/button-subtle-ai/button-subtle-ai.js",
+    "./components/button-subtle-ai.js": "./src/components/button-ai/button-subtle-ai.js",
     "./components/card-overlay.js": "./src/components/card-overlay/card-overlay.js",
     "./components/checkbox-drawer.js": "./src/components/checkbox-drawer/checkbox-drawer.js",
     "./components/community-button.js": "./src/components/community/community-button.js",
